backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
supabase>=2.3.4
pydantic>=2.6.1
python-jose[cryptography]==3.3.0
httpx>=0.27.0
